requirements.txt
# Core Dependencies

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
tqdm

# Optimization & Scientific Computing
scipy>=1.11.0

# Validation
pydantic>=2.0.0

# NLP & Text Processing
nltk>=3.8.0
textblob>=0.17.0
spacy>=3.5.0

# LLM & Transformers
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=4.25.0

# Hugging Face Integration
huggingface-hub>=0.19.0

# Dashboard & UI
streamlit>=1.28.0
gradio>=4.0.0
plotly>=5.18.0

# Export & Visualization
xlsxwriter>=3.1.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
openai>=1.0.0

# OCR Libraries
pytesseract>=0.3.10
easyocr>=1.7.0
Pillow>=10.0.0